requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.4

# Configuration
django-environ==0.4.1



# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.29.0


# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.6.0
redis>=2.10.5
django-grappelli==2.9.1
django-teryt-tree==0.11.0
django-autofixture==0.12.1
django-leaflet==0.19.0
django-geojson==2.9.1
https://github.com/ad-m/django-geojson/archive/python35-support-rebased.zip # See https://github.com/makinacorpus/django-geojson/issues/82 and https://github.com/makinacorpus/django-geojson/pull/81
# Your custom requirements go here
djmail==1.0.0
django-atom==0.12.7
django-haystack==2.5.1
django-tinymce==2.4.0
